refactor(frontend): isolate the authentication logic in a custom hook
