+++ frontend/src/modules/auth/types.ts
@@ -1,4 +1,6 @@
 export type Session = {
+  session: string;
   userId: string;
   token: string;
+  validatingSession: boolean;
 };
